requirements.txt
# Web scraping
requests>=2.31.0
requests[socks]>=2.31.0
beautifulsoup4>=4.12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter (optional for production)

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
anyio>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# TOR support
stem>=1.8.2

# Performance
uvloop>=0.19.0
orjson>=3.9.0

# Caching & Background Tasks
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0  # Celery monitoring

# WebSocket support
python-socketio>=5.11.0
websockets>=12.0

# Notifications
python-telegram-bot>=20.7
aiosmtplib>=3.0.0  # Async SMTP client

# Rate Limiting (Optional - comment out if not using)
# slowapi>=0.1.9
# limits>=3.7.0